requirements.txt
cvxpy==1.6.0
datasets==3.2.0
lightning==2.5.0.post0
lm_eval==0.3.0
matplotlib==3.5.1
numpy==1.21.5
opt_einsum==3.3.0
psutil==5.9.0
PyYAML==6.0.2
scipy==1.8.0
sympy==1.13.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
